Ubacene opce funkcije i prvobitno testiranje

+++ src/ba/unsa/etf/rpr/Artikl.java
@@ -3,11 +3,11 @@ package ba.unsa.etf.rpr;
 public class Artikl {
 
     private String naziv="";
-    private double cijena=0;
+    private int cijena=0;
     private String kod="";
 
 
-    public Artikl(String naziv,double cijena,String kod){
+    public Artikl(String naziv,int cijena,String kod){
 
         this.naziv=naziv;
         this.cijena=cijena;
@@ -19,7 +19,7 @@ public class Artikl {
         return naziv;
     }
 
-    public double getCijena(){
+    public int getCijena(){
         return cijena;
     }
     public String getKod(){
